backend/requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
langgraph
langchain-text-splitters
pypdf
python-dotenv
chromadb
fastapi
uvicorn
certifi
openai
httpx
pyjwt
email-validator
supabase
